reorganized routers
